Expense_tracker/expense_tracker/requirements.txt
<<<<<<< HEAD
Django>=5.2.6
djangorestframework>=3.15
Pillow>=10.0.0
=======
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Django==5.2.6
django-crontab==0.7.1
django-seed==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
Faker==37.8.0
fastapi==0.116.1
Flask==3.1.1
fonttools==4.61.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.7
mysqlclient==2.2.7
numpy==2.3.5
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pywebio==1.8.4
reportlab==4.4.5
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.3
toposort==1.10
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.35.0
Werkzeug==3.1.3
>>>>>>> 1da35822626378c190b037b04eb757b0426bd0bf
